@@ -1,88 +1,85 @@
 {
-    "name": "arroyo",
-    "version": "1.0.0",
-    "license": "MPL-2.0",
-    "engines": {
-      "node": "16",
-      "yarn": "1"
-    },
-    "dependencies": {
-      "@emotion/core": "10.0.35",
-      "@octokit/rest": "18.12.0",
-      "@react-three/drei": "^9.6.2",
-      "@react-three/fiber": "^8.0.12",
-      "@react-three/postprocessing": "^2.3.2",
-      "@testing-library/dom": "^8.13.0",
-      "@testing-library/jest-dom": "5.16.4",
-      "@testing-library/react": "13.1.1",
-      "@testing-library/user-event": "14.1.1",
-      "@tweenjs/tween.js": "18.6.4",
-      "dompurify": "^2.3.6",
-      "framer-motion": "6.3.1",
-      "glsl-easings": "^1.0.0",
-      "glsl-noise": "^0.0.0",
-      "glslify-loader": "^2.0.0",
-      "marked": "^4.0.14",
-      "raw-loader": "^4.0.2",
-      "react": "18.0.0",
-      "react-animations": "^1.0.0",
-      "react-dom": "18.0.0",
-      "react-intersection-observer": "9.0.0",
-      "react-scripts": "5.0.1",
-      "react-scroll": "1.8.7",
-      "three": "0.139.2",
-      "typed.js": "2.0.12",
-      "typescript": "4.6.3",
-      "web-vitals": "2.1.4"
-    },
-    "scripts": {
-      "analyze": "bundle-analyzer 'dist/static/js'",
-      "start": "react-app-rewired start",
-      "build": "react-app-rewired build && rm -rf dist && mv build dist",
-      "test": "react-app-rewired test",
-      "eject": "react-scripts eject",
-      "postinstall": "husky install"
-    },
-    "lint-staged": {
-      "*.**": "pretty-quick --staged"
-    },
-    "eslintConfig": {
-      "extends": [
-        "react-app",
-        "react-app/jest"
-      ]
-    },
-    "browserslist": {
-      "production": [
-        ">0.2%",
-        "not dead",
-        "not op_mini all"
-      ],
-      "development": [
-        "last 1 chrome version",
-        "last 1 firefox version",
-        "last 1 safari version"
-      ]
-    },
-    "devDependencies": {
-      "@octokit/types": "^6.34.0",
-      "@types/dat.gui": "0.7.7",
-      "@types/dompurify": "^2.3.3",
-      "@types/jest": "27.4.1",
-      "@types/marked": "^4.0.3",
-      "@types/node": "17.0.5",
-      "@types/react": "17.0.43",
-      "@types/react-dom": "18.0.2",
-      "@types/react-scroll": "^1.8.2",
-      "@types/three": "^0.139.0",
-      "bundle-analyzer": "^0.0.6",
-      "husky": "^5.2.0",
-      "lint-staged": "^10.5.4",
-      "prettier": "2.2.1",
-      "pretty-quick": "^3.1.0",
-      "react-app-rewired": "^2.2.1",
-      "yarn-audit-fix": "^9.2.2"
-    },
-    "packageManager": "yarn@1.22.18"
-  }
-  
+  "name": "arroyo",
+  "version": "1.0.0",
+  "license": "MPL-2.0",
+  "engines": {
+    "node": "16",
+    "yarn": "1"
+  },
+  "dependencies": {
+    "@emotion/core": "10.0.35",
+    "@octokit/rest": "18.12.0",
+    "@react-three/drei": "^9.56.27",
+    "@react-three/fiber": "^8.11.5",
+    "@react-three/postprocessing": "^2.7.0",
+    "@testing-library/dom": "^8.13.0",
+    "@testing-library/jest-dom": "5.16.4",
+    "@testing-library/react": "13.1.1",
+    "@testing-library/user-event": "14.1.1",
+    "@tweenjs/tween.js": "18.6.4",
+    "dompurify": "^2.3.6",
+    "framer-motion": "6.3.1",
+    "glsl-easings": "^1.0.0",
+    "glsl-noise": "^0.0.0",
+    "marked": "^4.0.14",
+    "react": "18.0.0",
+    "react-animations": "^1.0.0",
+    "react-dom": "18.0.0",
+    "react-intersection-observer": "9.0.0",
+    "react-scripts": "5.0.1",
+    "react-scroll": "1.8.7",
+    "three": "0.150.0",
+    "typed.js": "2.0.12",
+    "typescript": "4.9.5",
+    "web-vitals": "2.1.4"
+  },
+  "scripts": {
+    "analyze": "bundle-analyzer 'dist/static/js'",
+    "start": "craco start",
+    "build": "craco build && rm -rf dist && mv build dist",
+    "test": "craco test",
+    "eject": "react-scripts eject",
+    "postinstall": "husky install"
+  },
+  "lint-staged": {
+    "*.**": "pretty-quick --staged"
+  },
+  "eslintConfig": {
+    "extends": [
+      "react-app",
+      "react-app/jest"
+    ]
+  },
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
+  },
+  "devDependencies": {
+    "@craco/craco": "^7.0.0",
+    "@octokit/types": "^6.34.0",
+    "@types/dat.gui": "0.7.7",
+    "@types/dompurify": "^2.3.3",
+    "@types/jest": "27.4.1",
+    "@types/marked": "^4.0.3",
+    "@types/node": "17.0.5",
+    "@types/react": "17.0.43",
+    "@types/react-dom": "18.0.2",
+    "@types/react-scroll": "^1.8.2",
+    "@types/three": "^0.149.0",
+    "bundle-analyzer": "^0.0.6",
+    "husky": "^5.2.0",
+    "lint-staged": "^10.5.4",
+    "prettier": "2.2.1",
+    "pretty-quick": "^3.1.0",
+    "yarn-audit-fix": "^9.3.9"
+  },
+  "packageManager": "yarn@1.22.18"
+}
